requirements.txt
openai>=1.3.0
anthropic>=0.8.0
pydantic>=2.0.0
prompt_toolkit>=3.0.0
rich>=13.0.0
python-dotenv>=1.0.0